fixed angle

+++ include/motion.h
@@ -66,7 +66,7 @@ void rotate_to(Vector target);
 
 /** Rotate the robot by given angle 
   * Angle should be in [-180, 180] **/
-void rotate(double);
+void rotate(int);
 
 
 #endif // MOTION_H
